True kerning and floating point precision
In UTF8 mode use true kerning, and use floating point subpixel precision to place characters

+++ ScriptParser.h
@@ -404,7 +404,7 @@ protected:
     bool legacy_english_mode; // 0 ... not english, 1 ... english and fixed-width and using `
     bool skip_newline_mode; // 0 if / is not detected at end of line; 1 if / is detected at end of line
     bool prev_skip_newline_mode;
-    int skip_newline_offset;
+    float skip_newline_offset;
     bool prev_atsign;
 #endif
